requirements.txt
flask
pillow
inference-sdk
python-dotenv